requirements.txt
asttokens==2.0.8
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
colorlog==6.7.0
cycler==0.11.0
decorator==5.1.1
executing==1.0.0
fonttools==4.37.1
idna==3.3
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
numpy==1.23.2
packaging==21.3
pandas==1.4.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.11.2
six==1.16.0
stack-data==0.5.0
threadpoolctl==3.1.0
torch-geometric==2.1.0.post1
torch-scatter==2.0.9
torch-sparse==0.6.15
torchmetrics==0.9.3
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
zenlog==1.1